Add 2 liks to users and events dashboards
