+++ client/src/components/Routes.jsx
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import { Route } from 'react-router-dom';
 import HomePage from './HomePage';
 import AdminPage from './AdminPage';
+import AdminPageUsers from './AdminPageUsers';
 import MapLayout from './MapLayout';
 import ConfirmEmail from './ConfirmEmail';
 import RecoverInfo from './RecoverInfo';
@@ -26,7 +27,8 @@ render(){
         <Route path="/" component={HomePage} exact />
         <Route path="/recover" component={HomePage} exact />
       <div className="container">
-        <Route path="/admin-page" component={AdminPage} />
+        <Route path="/admin-page" component={AdminPage} exact />
+        <Route path="/admin-page/users" component={AdminPageUsers} exact />
         <Route path="/email_confirm/:token" component={ConfirmEmail} />
         <Route path="/recover-info" component={RecoverInfo} />
         <Route path="/change-password/:token" component={ChangePassword} />
